+++ core-interface/CoreModules/async_thread.hh
@@ -1,31 +1,2 @@
-#pragma once
-#include "CoreModules/CoreProcessor.hh"
-#include "util/callable.hh"
-#include <memory>
-
-namespace MetaModule
-{
-
-class AsyncThread {
-public:
-	AsyncThread(CoreProcessor *module);
-	AsyncThread(CoreProcessor *module, Callback &&action);
-
-	void start();
-	void start(Callback &&action);
-
-	void stop();
-	void run_once();
-
-	bool is_enabled();
-
-	~AsyncThread();
-
-private:
-	Callback action{};
-
-	struct Internal;
-	std::unique_ptr<Internal> internal;
-};
-
-} // namespace MetaModule
+// #warning Deprecated: Please use `#include "threads/async_thread.hh"` instead of `#include "async_thread.hh"`
+#include "../threads/async_thread.hh"
